requirements.txt
timm==1.0.11
torch==2.3.1
torchvision==0.18.1
faiss-cpu==1.12.0
scikit-learn==1.5.0
numpy>=1.19.5,<2.0
Pillow==10.4.0
tqdm==4.66.4
opencv-python==4.9.0.80
matplotlib==3.8.4
faster-whisper==1.2.0
ctranslate2>=4.0,<=4.4.0